Enable running of multiple test setups against the same tests in datamodel test runner
